Load the People fall step as a float
Declare the shared 40.0 step constant with its defining type so the fall-motion accumulator uses the retail floating-point ABI instead of converting the raw bit pattern from u32.

Constraint: Keep the defining sdata2 symbol and split unchanged.

Confidence: high

Scope-risk: narrow

Tested: ninja -j4 build/GC6E01/src/game/people/people.o build/GC6E01/report.json; fn_801821B8 fuzzy 65.76413% -> 65.79783%.

+++ src/game/people/people.c
@@ -1424,7 +1424,7 @@ void fn_80181EB0(u32 groupId, u32 index) {
 #endif
 
 /* 0x801821B8 | 0xE60 */
-extern u32 lbl_8047D7D8;
+extern f32 lbl_8047D7D8;
 extern f32 lbl_8047D7A0;
 extern f32 lbl_8047D7A4;
 extern u32 lbl_8047D7D0;
